fix DriveModelTest
tuta#3885
