+++ src/applications/drive-app/driveLocator.ts
@@ -266,9 +266,7 @@ class DriveLocator implements CommonLocator {
 			this.logins,
 			this.userManagementFacade,
 			EnvProvider.get().isDesktop() ? new WebFileResolver(window.nativeApp, this.fileApp, this.desktopSystemFacade) : null,
-			windowFacade,
 			redraw,
-			() => showWindowCloseConfirmation("closeWindowWithActiveTransfers_msg"),
 			this.connectivityModel,
 			this.searchModel,
 			this.searchRouter,
