+++ src/applications/mail-app/mailLocator.ts
@@ -1467,9 +1467,7 @@ class MailLocator implements CommonLocator {
 			this.logins,
 			this.userManagementFacade,
 			EnvProvider.get().isDesktop() ? new WebFileResolver(window.nativeApp, this.fileApp, this.desktopSystemFacade) : null,
-			windowFacade,
 			redraw,
-			() => showWindowCloseConfirmation("closeWindowWithActiveTransfers_msg"),
 			this.connectivityModel,
 			this.driveSearchModel,
 			await this.unscopedSearchRouter(),
